No Svelte!
Not quite sold on Svelte yet, it doesn't fully work nicely when using it with TypeScript. I do really like it's component model, but the DX isn't as nice to set up for tooling. I'm probably going to fold on this down the line anyways, so this is the start to going back and forth on a great technology that I just don't understand how to use properly yet XD